✨ feat(home_history_and_submission.dart): add EmptyHistoryAndSubmission widget This commit adds a new widget called EmptyHistoryAndSubmission to the home_history_and_submission.dart file. The widget is used to display a message and an animation when there is no data to display. It also includes a button to refresh the data. The widget takes in a HomeController, a text string, and an optional onPressed function.
✨ feat(home_history_and_submission.dart): add ListRiwayatInput widget
This commit adds the `ListRiwayatInput` widget to the `home_history_and_submission.dart` file. The widget is used to display a list of input history for a user. The list is scrollable and can be refreshed by pulling down. The list items are clickable and navigate to a new page when tapped. The widget uses the `HomeController` to get the list of input history and display it.

✨ feat(home_history_and_submission.dart): add list of submission history
This commit adds a new file, `list_riwayat_pengajuan.dart`, which contains a widget that displays a list of submission history. The list is scrollable and can be refreshed by pulling down. Each item in the list displays the name of the borrower, submission ID, submission date, and status. The status is displayed as a chip with different colors depending on the status. The widget also allows the user to navigate to the submission detail page by tapping on an item.

🔨 refactor(home_history_and_submission.dart): extract empty history widget to a separate file
The code for displaying an empty history message was extracted to a separate widget to improve code organization and readability. The new widget is called EmptyHistoryAndSubmission and takes in a text and an onPressed function as parameters. The HomeHistoryAndSubmission widget now uses this new widget to display the empty history message.

🎨 style(home_history_and_submission.dart): refactor widget to use EmptyHistoryAndSubmission
This commit refactors the HomeHistoryAndSubmission widget to use the newly created EmptyHistoryAndSubmission widget. This improves code readability and reduces code duplication. The EmptyHistoryAndSubmission widget is now used to display a message when there is no data available.

🔥 refactor(home_history_and_submission.dart): remove unused classes
The classes `ListRiwayatInput`, `ListRiwayatPengajuan`, and `TitlePengajuan` are no longer used in the codebase and have been removed. This improves code readability and maintainability.
